# Optimize querying for missing albums